上传单项

+++ app/src/main/java/com/coahr/thoughtrui/mvp/model/Bean/Attendance.java
@@ -76,7 +76,7 @@ public class Attendance {
         private String code;
         private int count;
         private String dealerId;
-        private int endTime;
+        private long endTime;
         private double latitude;
         private String location;
         private double longitude;
@@ -173,11 +173,11 @@ public class Attendance {
             this.dealerId = dealerId;
         }
 
-        public int getEndTime() {
+        public long getEndTime() {
             return endTime;
         }
 
-        public void setEndTime(int endTime) {
+        public void setEndTime(long endTime) {
             this.endTime = endTime;
         }
